lead_scraper/requirements.txt
requests==2.32.3
openai==2.26.0
apify-client==1.8.1
pyairtable==3.2.0
python-dotenv==1.0.1
tenacity==9.1.4
